Make API be contained within V1 module and make default URL point to V1 using ApiConstraints
