--- BookRater/app/controllers/api/reviews_controller.rb
@@ -1,17 +1,0 @@
-module Api
-	class ReviewsController < ApplicationController
-		def index
-	 		render json: Review.where(book_id: params[:book_id])
-		end
-
-		def create
-			# @review = Review.new(review_params)
-			# @book = Book.find(params[:book_id])
-		end
-
-		def update
-			# @book = Book.find(params[:book_id])
-			# @review = Review.find(params[:id])
-		end
-	end
-end
